% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copenhagen.R
\docType{data}
\name{copenhagen}
\alias{copenhagen}
\title{The Copenhagen Network}
\format{
'copenhagen' is a List of three objects. 
\enumerate{
\item \strong{edgelist}: a List of three \code{data.frame} objects: the network of physical proximity ('proximity'), the network of phone calls ('calls'), the network of text messages ('sms').

\item \strong{actors}: a vector of length 800 that contains all the actors ID's.

\item \strong{attributes}: a List of two \code{data.frame} objects: 'fb_friends' and 'gender'.
\describe{
\item{fb_friends}{\code{data.frame} with two columns ('actor1', 'actor2') describing the two actors that are friends on Facebook. Only the edges describing the friendships are recorded (If an edge is not included that means no friendship). Such friendships were formed before the end of the observation and didn't dissolve until after the end of the observation period.}
\item{gender}{\code{data.frame} with 2 columns 'actors' and 'female'. The latter column assumes value 1 if actor is female, 0 otherwise. The gender for some actors is not available.}
}
}
}
\source{
\url{https://figshare.com/articles/dataset/The_Copenhagen_Networks_Study_interaction_data/7267433/1}
}
\usage{
data(copenhagen)
}
\description{
The data represent a multi-layer temporal network that connects more than 700 university students at the Technical University of Denmark over a period of four weeks. 
 The temporal network includes: the network of physical proximity among the participants, the network of phone calls, the network of text messages. Additional information about Facebook friendships and gender of actors are available.
}
\section{Network of physical proximity (undirected, interval timing)}{

Edgelist with 2426279 events occurring among 706 actors and 4 columns describing the sequence of physical proximity events.
\describe{
 \item{time}{Numeric value indicating the seconds transpired since the onset. Events are recorded every 300 seconds and the starting time value is 0.}
 \item{actor1}{ID (integer value) of the first actor physically close to the second.}
 \item{actor2}{ID (integer value) of the second actor physically close to the first.}
 \item{rssi}{Received Signal Strength (RSSI) of the device used to measure the physical proximity: when two actors are physically close the RSSI increases, otherwise when the value of the RSSI decreases this indicates that actors are distant or that there are obstacles in between.}
}
}

\section{Network of phone calls (directed, interval timing)}{

Edgelist with 3600 calls occurring among 540 actors and 4 columns describing the sequence of calls.
\describe{
 \item{time}{Numeric value indicating the seconds transpired since the onset. The starting time value is 0.}
 \item{actor1}{ID (integer value) of actor starting the call.}
 \item{actor2}{ID (integer value) of the actor receiving the call.}
 \item{duration}{duration of the call. A missed call is recorded with duration = -1.}
}
}

\section{Network of text messages (directed, interval timing)}{

Edgelist with 24333 text messages sent among 800 actors and 3 columns describing the sequence of messages (message content not available). 
\describe{
 \item{time}{Numeric value indicating the seconds transpired since the onset. The starting time value is 0.}
 \item{actor1}{ID (integer value) of actor sending the message.}
 \item{actor2}{ID (integer value) of the actor receiving the message.}
}
}

\references{
Sapiezynski, P., Stopczynski, A., Lassen, D. D., & Lehmann, S. (2019). Interaction data from the Copenhagen Networks Study. Scientific Data, 6(1), 1–10. \url{https://doi.org/10.1038/s41597-019-0325-x}
}
\keyword{datasets}
