% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disasters_bycountry.R
\name{disasters_bycountry}
\alias{disasters_bycountry}
\title{disasters_bycountry function}
\usage{
disasters_bycountry(exactcountry = "Bangladesh")
}
\arguments{
\item{exactcountry}{Defaults to "Bangladesh"
disasters_bycountry()}
}
\description{
disasters_bycountry function
}
