% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highschool2011.R
\docType{data}
\name{highschool2011}
\alias{highschool2011}
\title{Contacts between high school students.}
\format{
List with elements 'edgelist' and 'attributes'.
}
\source{
\url{https://sociopatterns.org/datasets/high-school-dynamic-contact-networks/}
}
\usage{
data(highschool2011)
}
\description{
This dataset contains the sequence of 10685 face-to-face contacts between 
high school students and their teachers in Marseilles, France during 4 days 
in December 2011. The data collection involved 118 students of three 
different classes and 8 teachers. Wearable proximity sensors recorded with a 
20-second time interval every face-to-face contact between individuals that 
was active in the last 20 second window. A contact between two individuals 
is defined as the set of successive time-windows of 20 seconds during which 
the individuals are detected in contact. In the 'edgelist'  element, these 
records are collected in a relational event sequence with for every contact 
between a pair of individuals the time of onset and its duration. Time is 
measured in seconds. The events in the sequence are undirected. Multiple 
events can occur at the same time point and the individuals can be in 
contact with multiple other individuals at the same time. Information on the 
gender and class/role of the individuals is available in the 'attributes' 
element.
}
\section{edgelist}{

Dataframe with 10685 rows and 4 columns that contains the relational event 
sequence with face-to-face contacts between the students and teachers. 
\describe{
 \item{time}{Numeric value that indicates the onset of the contact in 
seconds.}
 \item{actor1}{Integer value with the ID of the first actor in the contact.}
 \item{actor2}{Integer value with the ID of the second actor in the contact.}
 \item{duration}{Numeric value that indicates the duration of the contact in 
seconds.}
}
}

\section{attributes}{

Dataframe with 126 rows and 4 columns describing the gender and class/role 
of the students and teachers.
\describe{
 \item{id}{Integer value with the ID of the individual.}
 \item{time}{Change time of the attributes. Since the attributes are 
constant over the course of the study period, this value is equal to 0 for 
all entries.}
 \item{class}{String with the class/role of the individual ('PC' stands for 
classes that focus on physics and chemistry and 'PSI' stands for classes 
that focus on engineering studies).}
 \item{gender}{String with the gender of the individual ('M' = male, 'F' = 
female).}
}
}

\references{
Fournet & Barrat (2014) PloS one 9(9) e107878 \doi{10.1371/journal.pone.0107878}
}
\seealso{
\link{highschool2012}
}
\keyword{datasets}
