% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highschool2013.R
\docType{data}
\name{highschool2013}
\alias{highschool2013}
\title{Contacts between high school students.}
\format{
List with elements 'edgelist', 'attributes', 'contacts',
'friendship' and 'facebook'.
}
\source{
\url{https://sociopatterns.org/datasets/high-school-contact-and-friendship-networks/}
}
\usage{
data(highschool2013)
}
\description{
This dataset contains the sequence of 67,613 face-to-face contacts between
high school students and their teachers in Marseilles, France from December
2nd until December 6th, 2013. The data collection involved 329 students of
nine different classes. Wearable proximity sensors recorded with a 20-second
time interval every face-to-face contact between students that was active in
the last 20 second window. A contact between two students is defined as the
set of successive time-windows of 20 seconds during which the students are
detected in contact. In the 'edgelist'  element, these records are collected
in a relational event sequence with for every contact between a pair of
students the time of onset and its duration. Time is measured in seconds.
The events in the sequence are undirected. Multiple events can occur at the
same time point and the students can be in contact with multiple other
students at the same time. On the 4th day of the study, students were asked
to report the list of other students they had contact with during the day in
the high school and give the approximate aggregated duration of these
contacts. In total, 119 students reported on their contacts. The resulting
directed reported contact network is available in the 'contacts' element.
Moreover, in total 133 students reported the name of their friends in the
high school. The resulting directed friendship network is available in the
'friendship' element. Information on whether or not the students are linked
on Facebook is available in the 'facebook' element. Information on the
gender and class of the students is available in the 'attributes' element.
}
\section{edgelist}{

Dataframe with 67613 rows and 4 columns that contains the relational event
sequence with face-to-face contacts between the students.
\describe{
 \item{time}{Numeric value that indicates the timestamp for the onset of the
contact expressed in UNIX ctime.}
 \item{actor1}{Integer value with the ID of the first actor in the contact.}
 \item{actor2}{Integer value with the ID of the second actor in the contact.}
 \item{duration}{Numeric value that indicates the duration of the contact in
seconds.}
}
}

\section{attributes}{

Dataframe with 329 rows and 4 columns describing the gender and class of the
students.
\describe{
 \item{id}{Integer value with the ID of the individual.}
 \item{time}{Change time of the attributes. Since the attributes are
constant over the course of the study period, this value is equal to 0 for
all entries.}
 \item{class}{String with the class of the individual.}
 \item{gender}{String with the gender of the individual ('M' = male, 'F' =
female). For 7 students, the gender is missing.}
}
}

\section{contacts}{

Matrix with the directed reported contacts network; 119 students reported on
the duration of their contacts with other students in the high school on the
4th day of the study period (1 = at most 5 minutes, 2 = between 5 and 15
minutes, 3 = between 15 minutes and 1 hour, 4 = more than one hour).
Row names and column names correspond to the respective actors.
}

\section{friendship}{

Matrix with the directed reported friendships network; 133 reported on their
friends (1 = did report a friendship, 0 = did not report a friendship). Row
names and column names correspond to the respective actors.
}

\section{facebook}{

Matrix with the undirected Facebook links (1 = are friends on Facebook, 0 =
are not). Row names and column names correspond to the respective actors.
}

\references{
Mastrandrea et al. (2015) PloS one 10(9) e0136497 \doi{10.1371/journal.pone.0136497}
}
\seealso{
\link{highschool2011}, \link{highschool2012}
}
\keyword{datasets}
