% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICCSS17.R
\docType{data}
\name{ICCSS17}
\alias{ICCSS17}
\title{International Conference on Computer Social Science 2017 (ICCSS17)}
\format{
List with elements 'edgelist' and 'attributes'
}
\source{
\url{https://zenodo.org/record/2531537#.YPEwf-gzaUl}
}
\usage{
data(ICCSS17)
}
\description{
This dataset contains the sequence of 70910 face-to-face contacts between
visitors of the International Conference on Computer Social Science 2017
(ICCSS17) in Cologne, July 10 to 13, 2017. The data collection involved 262
of the 339 conference attendees. Wearable proximity sensors recorded every
face-to-face contact between individuals (physical proximity within
approximately 1.5m) with a 20-second time resolution. A contact between two
individuals is defined as the set of successive time-windows of 20-second
during which the individuals are detected in close proximity. In the
'edgelist' element, these records are collected in a relational event
sequence with for every contact between a pair of individuals the time of
onset and its duration. Time is measured in seconds. The events in the
sequence are undirected. Multiple events can occur at the same time point
and the individuals can be in contact with multiple other individuals at the
same time. Socio-demographic information about the participants (seniority,
disciplinary background, country of residence, mother tongue, age, gender,
role at conference) was collected through a survey and available in the
'attributes' element.
}
\section{edgelist}{

Dataframe with 70910 rows and 4 columns that contains the relational event
sequence with face-to-face contacts between the participants.
\describe{
 \item{time}{POSIXct value that indicates the onset of the contact.}
 \item{actor1}{Integer value with the ID of the first actor in the contact.}
 \item{actor2}{Integer value with the ID of the second actor in the contact.}
 \item{duration}{Numeric value that indicates the duration of the contact in
seconds.}
}
}

\section{attributes}{

Dataframe with 262 rows and 9 columns describing the socio-demographic
information about the participants.
\describe{
 \item{id}{Integer value with the ID of the individual.}
 \item{time}{Change time of the attributes. Since the attributes are
constant over the course of the study period, this value is equal to 0 for
all entries.}
 \item{age}{Age range for the individual. One of 'younger than 30 years',
'30 to 39 years', '40 to 49 years', or '50 years and older'.}
 \item{gender}{Gender of the individual, 'Female' or 'Male'.}
 \item{country}{Code for the country of residence of the individual.}
 \item{language}{Code for the mother tongue of the individual.}
 \item{seniority}{Academic seniority of the individual at the time of the
conference. One of 'Bachelor/Master student', 'PhD student',
'Postdoctoral researcher', 'Junior/Assistant professor', 'Associate/Full
professor', 'Other'}
 \item{background}{Academic background of the individual. One of 'Social/
Political/Behavioral Sciences', 'Computer/Information Science', 'Math/
Physics/Biology', 'Media/Communication/Linguistics', or 'Others'. }
 \item{role}{Role in the conference of the individual. One of 'Speaker',
'Poster presenter', 'Participation only', 'Staff'}
 \item{previous}{Did the individual participate to a previous edition? (Yes
or No).}
}
}

\references{
Génois, Zens, Lechner, Rammstedt & Strohmaier (2019). Building
connections: How scientists meet each other during a conference.
\url{https://arxiv.org/pdf/1901.01182.pdf}
}
\keyword{datasets}
