% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectious.R
\docType{data}
\name{infectious}
\alias{infectious}
\title{Contacts between visitors of the artscience exhibition 'INFECTIOUS: STAY 
AWAY' on May 21st, 2009.}
\format{
Dataframe with 1332 rows and 4 columns that contains the relational 
event sequence with face-to-face contacts between the visitors. 
\describe{
 \item{time}{Numeric value indicating the timestamp for the onset of the 
contact expressed in UNIX ctime.}
 \item{actor1}{Integer value with the ID of the first actor in the contact.}
 \item{actor2}{Integer value with the ID of the second actor in the contact.}
 \item{duration}{Numeric value indicating the duration of the contact in 
seconds.}
}
}
\source{
\url{https://sociopatterns.org/datasets/infectious-sociopatterns-dynamic-contact-networks/}
}
\usage{
data(infectious)
}
\description{
This dataset contains the sequence of 1332 face-to-face contacts between 96 
visitors of the artscience exhibition 'INFECTIOUS: STAY AWAY' on May 21st, 
2009. Wearable proximity sensors recorded with a 20-second time interval 
every face-to-face contact between visitors that was active in the last 20 
second window. A contact between two individuals is defined as the set of 
successive time-windows of 20 seconds during which the individuals are 
detected in contact. In the 'infectious' edgelist, these records are 
collected in a relational event sequence with for every contact between a 
pair of individuals the time of onset and its duration. Time of the events 
is expressed in UNIX ctime. The events in the sequence are undirected. 
Multiple events can occur at the same time point and the individuals can be 
in contact with multiple other individuals at the same time.
}
\references{
Isella et al. (2011) Journal of Theoretical Biology 271, 166 \doi{10.1016/j.jtbi.2010.11.033}
}
\keyword{datasets}
