% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\docType{data}
\name{info2}
\alias{info2}
\title{Exogenous information of 5 actors}
\format{
A dataframe with 7 rows and 4 variables:
\describe{
 \item{name}{numeric id of the actor}
 \item{time}{numeric value, describes when the value of the covariate
  changes, if it changes}
 \item{age}{age of the actor}
 \item{sex}{dichotomized sex of the actor (e.g., 0 = male, 1 = female)}
}
}
\usage{
data(info2)
}
\description{
A dataset containing the (hypothetical) age of the actors in a social
network of a relational event history \code{\link{randomREH2}}.
}
\seealso{
\code{\link{randomREH2}} for the relational event history.
}
\keyword{dataset}
