% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linux_replies.R
\docType{data}
\name{linux_replies}
\alias{linux_replies}
\title{Linux Kernel mailing list 2006-2013}
\format{
Dataframe with 1064718 rows and 4 columns. Each row represents a
reply from one member to another member of the mailing list.
\describe{
  \item{time}{Unix timestamp for the reply.}
  \item{actor1}{ID of the member that replies to another member's message,
  i.e., the 'source' or 'sender' of the message.}
  \item{actor2}{ID of the member to whose message is replied to,
  i.e., the 'target' or 'receiver' of the message.}
  \item{date}{POSIXct date and time belonging to the Unix timestamp.}
}
}
\source{
\url{https://networkrepository.com/comm-linux-kernel-reply.php},
\url{https://lkml.org/}
}
\usage{
data(linux_replies)
}
\description{
The Linux Kernel mailing list is a list where discussions on the design of,
and bugs in the Linux kernel take place. This data set contains the
time-stamped replies from members of the list to other members' messages
between 2006 and 2013. Note: this data set contains self-to-self edges.
}
\references{
Rossi, R. A., & Ahmed, N. K. (2015) The network data repository
with interactive graph analytics and visualization. AAAI.
\url{https://networkrepository.com}
}
\keyword{datasets}
