% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathoverflow.R
\docType{data}
\name{mathoverflow}
\alias{mathoverflow}
\title{MathOverflow Network}
\format{
'mathoverflow' is a List of two objects. 
\enumerate{
\item \strong{edgelist}: a List containing a \code{data.frame} describing a network of 390441 events among 24759 actors (users in the website)

\describe{
 \item{time}{timestamp of the interaction (interval timing).}
 \item{actor1}{ID (integer value) of the actor that writes an answer/comment.}
 \item{actor2}{ID (integer value) of the actor that receives the answer/comment.}
}

\item \strong{actors}: a vector of actors ID's.
}
}
\source{
\url{http://konect.cc/networks/sx-mathoverflow/}
}
\usage{
data(mathoverflow)
}
\description{
This data contain interactions from the StackExchange site "MathOverflow". In this temporal network, directed edges between users represent three types of interactions: answering a question of another user, commenting on another user's question, and commenting on another user's answer.
}
\references{
Jure Leskovec. Stanford Network Analysis Project. \url{http://snap.stanford.edu/}, September 2014.
}
\keyword{datasets}
