% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryschool.R
\docType{data}
\name{primaryschool}
\alias{primaryschool}
\title{Contacts in a primary school.}
\format{
List with elements 'edgelist' and 'attributes'.
}
\source{
\url{https://sociopatterns.org/datasets/primary-school-temporal-network-data/}
}
\usage{
data(primaryschool)
}
\description{
This dataset contains the sequence of 77521 face-to-face contacts between 
primary school students and their teachers in Lyon, France on Thursday 
October 1st and Friday October 2nd 2009. The data collection involved 232 
students of five different grades of each two classes (i.e., 10 classes in 
total) and 10 teachers. Wearable proximity sensors recorded with a 20-second 
time interval every face-to-face contact between individuals that was active 
in the last 20 second window. A contact between two individuals is defined 
as the set of successive time-windows of 20 seconds during which the 
individuals are detected in contact. In the 'edgelist' element, these 
records are collected in a relational event sequence with for every contact 
between a pair of individuals the time of onset and its duration. Time is 
measured in seconds. The events in the sequence are undirected. Multiple 
events can occur at the same time point and the students can be in contact 
with multiple other students at the same time. Information on the gender and 
class/role of the students and teachers is available in the 'attributes' 
element.
}
\section{edgelist}{

Dataframe with 77521 rows and 4 columns that contains the relational event 
sequence with face-to-face contacts between the students. 
\describe{
 \item{time}{Numeric value indicating the timestamp for the onset of the 
contact in seconds since onset of the study period.}
 \item{actor1}{Integer value with the ID of the first actor in the contact.}
 \item{actor2}{Integer value with the ID of the second actor in the contact.}
 \item{duration}{Numeric value indicating the duration of the contact in 
seconds.}
}
}

\section{attributes}{

Dataframe with 242 rows and 4 columns describing the gender and class/role 
of the students and teachers. 
\describe{
 \item{id}{Integer value with the ID of the individual.}
 \item{time}{Change time of the attributes. Since the attributes are 
constant over the course of the study period, this value is equal to 0 for 
all entries.}
 \item{class}{String with the grade and class/role of the students and 
teachers.}
 \item{gender}{String with the gender of the student ('M' = male, 'F' = 
female). Note that for 15 individuals, the gender is unknown.}
}
}

\references{
Gemmetto et al. (2014) BMC Infectious Diseases 14, 695 \doi{10.1186/s12879-014-0695-9}, Stehle et al. (2011) PloS one 6(8) e23176 \doi{10.1371/journal.pone.0023176}
}
\keyword{datasets}
