% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\docType{data}
\name{randomREH2}
\alias{randomREH2}
\title{Generated Relational Event History with Duration, Type, and Weight}
\format{
\code{data(randomREH2)} will load a list containing following objects:
\describe{
   \item{\code{edgelist}}{
   a \code{data.frame} that contains the random sequence of events. Columns of the edgelist are:
           \describe{
               \item{\code{time}}{the timestamp indicating the time at which each event occurred;}
               \item{\code{actor1}}{the name of the actor that generated the relational event;}
               \item{\code{actor2}}{the name of the actor that received the relational event;}
               \item{\code{end}}{the time of ending the relational event;}
               \item{setting}{the setting (type) of the event, either \code{social} or \code{work};}
               \item{duration}{the duration of each event, which can also function as event weight for instance.}
           }
    }
}
}
\usage{
data(randomREH2)
}
\description{
A randomly generated sequence of relational events `randomREH2`, among 5
actors and 1000 events with duration. For starting an event, inertia and
reciprocity were used, and for (undirected) ending an event totaldegreeDyad
was used.
}
\seealso{
\code{\link{info2}} for actor covariates.
}
\keyword{datasets}
