% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travian.R
\docType{data}
\name{travian}
\alias{travian}
\title{Interactions between players of the multi-player online game Travian}
\format{
List with elements 'edgelist', 'actors' and 'attributes'
}
\source{
\url{http://ial.eecs.ucf.edu/travian.php}
}
\usage{
data(travian)
}
\description{
This dataset contains the sequence of 1,355,204 interactions between players
of the multi-player online game Travian between December 1st, 2009 and
December 31st, 2009. In the game, players have to develop their civilization
over a fixed period of time (a few months). In order to expand, they can
attack villages from other players, trade resources with other players and
form alliances from up to 60 players. Travian also has an in-game messaging
system through which players can send each other messages. The relational
event history in the 'edgelist' contains all attacks, trades and messages
between players during the study period. The 'attributes' elements contains
information on the alliances (communities) to which the players belong at
the start of each day.
}
\section{edgelist}{

Dataframe with 1,355,204 rows and 4 columns with the interactions between
the players.
\describe{
 \item{time}{POSIXct value with the timestamp for the interaction.}
 \item{actor1}{Integer value with the ID of the sender of the interaction.}
 \item{actor2}{Integer value with the ID of the receiver of the interaction.}
 \item{type}{String value with the type of the interaction (attack, trade or
message).}
}
}

\section{actors}{

Vector with the integer IDs of all 4,632 players that sent or received
interactions or were in the community structure during the study period.
}

\section{attributes}{

Dataframe with 138,960 rows and 3 columns describing for each day the
community to which the players belong.
\describe{
 \item{id}{Integer value with the ID of the player (corresponding to the ID
in 'edgelist').}
 \item{time}{POSIXct value corresponding to the start of the day.}
 \item{community}{Integer value with the ID of the community to which the
player belongs for the respective day. The communities are re-numbered at
each new day. The value '0' is reserved for players that were not in the
community structure for the respective day.}
}
}

\references{
Hajibagheri A., Sukthankar G., Lakkaraju K. (2016) A Holistic
Approach for Link Prediction in Multiplex Networks. In: Spiro E., Ahn YY.
(eds) Social Informatics. SocInfo 2016. Lecture Notes in Computer Science,
vol 10047. Springer, Cham. \url{https://doi.org/10.1007/978-3-319-47874-6_5},
Hajibagheri A., Sukthankar G., Lakkaraju K. (2017) Extracting Information
from Negative Interactions in Multiplex Networks Using Mutual Information.
In: Lee D., Lin YR., Osgood N., Thomson R. (eds) Social, Cultural, and
Behavioral Modeling. SBP-BRiMS 2017. Lecture Notes in Computer Science, vol
10354. Springer, Cham. \url{https://doi.org/10.1007/978-3-319-60240-0_39}
}
\keyword{datasets}
