% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikitalkbr.R
\docType{data}
\name{wikitalkbr}
\alias{wikitalkbr}
\title{Breton Wikipedia Talk Network}
\format{
'wikitalkbr' is a List of two objects.
\enumerate{
\item \strong{edgelist}: a List containing a \code{data.frame} describing a network of 10391 events among 1049 actors (users in the website)

\describe{
 \item{time}{timestamp of the interaction (interval timing).}
 \item{actor1}{ID (integer value) of the actor that writes an answer/comment.}
 \item{actor2}{ID (integer value) of the actor that receives the answer/comment.}
}

\item \strong{actors}: a vector of actors ID's.
}
}
\source{
\url{http://konect.cc/networks/wiki_talk_br/}
}
\usage{
data(wikitalkbr)
}
\description{
This data contain the communication network of the Breton Wikipedia. In this temporal network, a directed edges between two users mean that one user wrote a message on the talk page of another user (self-loops are excluded).
}
\references{
Jérôme Kunegis. KONECT – The Koblenz Network Collection. In Proc. Int. Conf. on World Wide Web Companion, pages 1343–1350, 2013. \url{https://dl.acm.org/doi/10.1145/2487788.2488173}
}
\keyword{datasets}
