% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WS16.R
\docType{data}
\name{WS16}
\alias{WS16}
\title{GESIS Computational Social Science Winter Symposium 2016 (WS16)}
\format{
List with elements 'edgelist' and 'attributes'
}
\source{
\url{https://psycharchives.org/en/item/26710d25-c387-4506-b5f5-60d1bd3009a5}
}
\usage{
data(WS16)
}
\description{
This dataset contains the sequence of 53164 face-to-face contacts between
visitors of the GESIS Computational Social Science Winter Symposium 2016
(WS16) in Cologne on November 30 and December 1, 2016. The data collection
involved 138 of the 149 conference attendees. Wearable proximity sensors
recorded every face-to-face contact between individuals (physical proximity
within approximately 1.5m) with a 20-second time resolution. A contact
between two individuals is defined as the set of successive time-windows of
20-second during which the individuals are detected in close proximity. In
the 'edgelist' element, these records are collected in a relational event
sequence with for every contact between a pair of individuals the time of
onset and its duration. Time is measured in seconds. The events in the
sequence are undirected. Multiple events can occur at the same time point
and the individuals can be in contact with multiple other individuals at the
same time. Socio-demographic information about the participants (seniority,
disciplinary background, country of residence, mother tongue, age, gender,
role at conference) was collected through a survey and available in the
'attributes' element.
}
\section{edgelist}{

Dataframe with 53164 rows and 4 columns that contains the relational event
sequence with face-to-face contacts between the participants.
\describe{
 \item{time}{POSIXct value that indicates the onset of the contact.}
 \item{actor1}{Integer value with the ID of the first actor in the contact.}
 \item{actor2}{Integer value with the ID of the second actor in the contact.}
 \item{duration}{Numeric value that indicates the duration of the contact in
seconds.}
}
}

\section{attributes}{

Dataframe with 138 rows and 9 columns describing the socio-demographic
information about the participants.
\describe{
 \item{name}{Integer value with the ID of the individual.}
 \item{time}{Change time of the attributes. Since the attributes are
constant over the course of the study period, this value is equal to 0 for
all entries.}
 \item{age}{Age range for the individual. One of 'younger than 30 years',
'30 to 39 years', or '40 years and older'.}
 \item{gender}{Gender of the individual, 'Female' or 'Male'.}
 \item{country}{Code for the country of residence of the individual.}
 \item{language}{Code for the mother tongue of the individual.}
 \item{seniority}{Academic seniority of the individual at the time of the
conference. One of 'Bachelor student', 'Master student', 'PhD student',
'Postdoctoral researcher', 'Assistant/associate professor', 'Full
professor', 'Other'}
 \item{background}{Academic background of the individual. One of 'Social
Science, Humanities, Political Science', 'Computer Science, Information
Science', 'Physics, Mathematics, Statistics',
'Media, Communication, Other'. }
 \item{role}{Role in the conference of the individual. One of 'Speaker',
'Poster presenter', 'Participation only', 'Staff'}
 \item{previous}{Did the individual participate to a previous edition? (Yes
or No).}
}
}

\references{
Genois, M., Zens, M., Oliveira, M., Lechner, C. M., Schaible, J.,
  & Strohmaier, M. (2023). Combining sensors and surveys to study social
  interactions: A case of four science conferences. \emph{Personality Science},
  4, Article e9957. \doi{10.5964/ps.9957}
}
\keyword{datasets}
