% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remdata-package.R
\docType{package}
\name{remdata-package}
\alias{remdata}
\alias{remdata-package}
\title{remdata: A collection of empirical and simulated relational event data sequences}
\description{
Empirical and simulated data for relational event analyses. Each dataset consists of an edgelist - the sequence events representing the link sent from one actor to another at a particular timepoint; and a dataframe describing the actors and their attributes if present.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TilburgNetworkGroup/remdata}
  \item Report bugs at \url{https://github.com/TilburgNetworkGroup/remdata/issues}
}

}
\author{
\strong{Maintainer}: Marlyne Meijerink-Bosman \email{m.l.meijerink@tilburguniversity.edu}

Authors:
\itemize{
  \item Diana Karimova \email{d.karimova@tilburguniversity.edu}
  \item Giuseppe Arena \email{g.arena@tilburguniversity.edu}
  \item Rumana Lakdawala \email{r.j.lakdawala@tilburguniversity.edu}
  \item Mahdi Shafiee Kamalabad \email{m.shafiee@tilburguniversity.edu}
  \item Fabio Generoso Vieira \email{f.v.generosovieira@tilburguniversity.edu}
  \item Roger Leenders \email{r.t.a.j.leenders@tilburguniversity.edu}
}

Other contributors:
\itemize{
  \item Joris Mulder \email{j.mulder3@tilburguniversity.edu} [contributor]
}

}
\keyword{internal}
